requirements.txt
aniso8601==9.0.1
appier==1.25.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
google==3.0.0
google-api==0.1.12
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.1
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2

python-dotenv~=0.19.2